added new classes according to homework #8 (Java 2 step. Part 3, 2/8 modificators)

+++ src/main/java/ru/itsjava/homeWork/oop/homeWork3/Figure.java
@@ -11,7 +11,7 @@ public class Figure {
     }
 
     //Методы
-    public int getPerimeter() {
+    public double getPerimeter() {
         System.out.println("Периметр внутри суперкласса Figure");
         return 0;
